feat: support OpenAPI v3
