+++ package.json
@@ -43,6 +43,7 @@
     "@babel/node": "^7.0.0",
     "@babel/preset-env": "^7.1.0",
     "@babel/register": "^7.0.0",
+    "ajv": "6.9.1",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.4",
     "chai": "^4.2.0",
@@ -60,7 +61,6 @@
     "webpack-cli": "^3.1.2"
   },
   "dependencies": {
-    "ajv": "^6.5.4",
     "commander": "^2.19.0",
     "errno": "0.1.7",
     "glob": "^7.1.3",
