+++ src/impl/v2/index.js
@@ -1,0 +1,11 @@
+// CONSTANTS
+
+const PATH__EXAMPLES = '$..examples.application/json';
+
+module.exports = {
+    getJsonPathToExamples
+};
+
+// IMPLEMENTATION DETAILS
+
+function getJsonPathToExamples() { return PATH__EXAMPLES; }
